14.54
added the ability to add more addresses to a contact in an arraylist. created a int to keep track of the size of it to assert later when adding multiple address. Added a JUnit test to make sure it was working correctly, and that the toString method will print out all of the email address on there own lines.